productiontaskqueue/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "productiontaskqueue"
version = "0.1.0"
description = "Production-grade task queue with retries, idempotency and crash safety"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Artinda Karakoc" }
]

dependencies = [
  "fastapi>=0.110",
  "uvicorn>=0.27",
  "psycopg2-binary>=2.9",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
